Feat: refresh하기 이전 state를 localstorage에 저장
