+++ src/component/presenter/layouts/User.jsx
@@ -18,6 +18,7 @@ const User = styled.div`
 
 function UserPage() {
   const { state } = useContext(MainContext);
+  console.log(state);
   if (state.curData.user) {
     return (
       <User>
