+++ src/context/actions.js
@@ -21,7 +21,9 @@ const FETCH_ADD_WORK = 'FETCH_ADD_WORK';
 const FETCH_CREATOR_WORKS = 'FETCH_CREATOR_WORKS';
 const FETCH_EDIT_WORK = 'FETCH_EDIT_WORK';
 const FETCH_SEARCH_DATA = 'FETCH_SEARCH_DATA';
+const GET_PRE_STATE = 'GET_PRE_STATE';
 export {
+  GET_PRE_STATE,
   FETCH_EDIT_WORK,
   FETCH_CREATOR_WORKS,
   FETCH_SEARCH_DATA,
